feat: organize cache classes
